Improved volume reading and response handling.  We now use a 16-bit volume range derived from a 12-bit ADC input, ensuring finer control over volume levels. The volume response can be toggled between linear and non-linear (logarithmic) modes for better auditory perception. Additionally, fixed function prototype for ReadVolume to return uint16_t.

+++ Application/CR2-VSCode/Core/Libraries/audio_engine.h
@@ -156,7 +156,7 @@ typedef struct {
 
 /* Hardware interface function pointer types */
 typedef void        ( *DAC_SwitchFunc )         ( GPIO_PinState setting );
-typedef uint8_t     ( *ReadVolumeFunc )         ( void );
+typedef uint16_t    ( *ReadVolumeFunc )         ( void );
 typedef void        ( *I2S_InitFunc )           ( void );
 
 extern DAC_SwitchFunc AudioEngine_DACSwitch;
